refactored data viz code into utility files
