+++ AnalyteExplorer/src/client/AnalyteExplorer/components/data_viz/LinePlot.tsx
@@ -1,6 +1,7 @@
 import React, { CSSProperties } from "react";
 import { D3LinePlot, D3LineData } from "./d3/LinePlot.d3";
 import LinePlotTooltip from "./LinePlotTooltip";
+import { setupLinePlotData } from "./dataVizHelperFuncs";
 import "./LinePlot.scss";
 
 export interface LinePlotProps {
@@ -24,32 +25,8 @@ const LinePlot: React.FC<LinePlotProps> = ({
     width: "100%",
   };
 
-  // transforms input data into format suitable for the line plot, also seperate data for trend line from cohorts
-  const setupData = (
-    data: Map<string, { x: number; y: number; study: string }[]>
-  ): D3LineData[][] => {
-    let cohortData: D3LineData[] = [];
-    let trendData: D3LineData[] = [];
-    for (const [cohort, linePoints] of data) {
-      if (linePoints.length > 0) {
-        const formattedData = {
-          name: cohort,
-          study: linePoints[0].study,
-          data: linePoints.sort((a, b) => a.x - b.x),
-        };
-        if (cohort === "Average") {
-          trendData.push(formattedData);
-        }
-        cohortData.push(formattedData);
-      }
-    }
-
-    // lineData contains trendData because it makes it easier to draw
-    return [cohortData, trendData];
-  };
-
   React.useEffect(() => {
-    const [formattedCohortData, formattedTrendData] = setupData(data);
+    const [formattedCohortData, formattedTrendData] = setupLinePlotData(data);
 
     D3LinePlot.create(
       name,
