+++ AnalyteExplorer/src/client/AnalyteExplorer/components/data_viz/dataVizConstants.ts
@@ -1,0 +1,10 @@
+export const tooltipStyle = {
+  position: "absolute",
+  backgroundColor: "white",
+  border: "1px solid black",
+  width: "auto",
+  minWidth: "200px",
+  height: "80px",
+  padding: "3px 3px",
+  display: "none",
+} as React.CSSProperties;
